backend/requirements.txt
fastapi==0.115.12
uvicorn==0.42.0
boto3==1.42.80
pydantic-settings==2.8.0
requests==2.33.1
python-multipart==0.0.22
pytest-cov==7.0.0
pytest==9.0.2
pytest-mock==3.14.0
SQLAlchemy==2.0.48
psycopg-binary==3.3.3
python-dotenv==1.1.0
loguru==0.7.3
mercadopago==2.3.0
httpx==0.28.1
sqlalchemy_filters==0.13.0

